+++ packages/frontend/editor-ui/src/app/components/MainHeader/ActionsDropdownMenu.vue
